refactor(portal): 移动组件位置
